🚚 Move all subdirectories from 'simnet/client' to 'simnet/client/components'
Moves the following subdirectories:
- account
- wish
- chronicle
