+++ simnet/client/components/chronicle/starrail.py
@@ -1,6 +1,6 @@
 from typing import Optional, Mapping, Dict, Any
 
-from simnet.client.chronicle.base import BaseChronicleClient
+from simnet.client.components.chronicle.base import BaseChronicleClient
 from simnet.errors import BadRequest, DataNotPublic
 from simnet.models.starrail.chronicle.characters import StarShipDetailCharacters
 from simnet.models.starrail.chronicle.notes import StarRailNote
