+++ simnet/client/components/wish/base.py
