+++ simnet/client/components/wish/genshin.py
@@ -1,7 +1,7 @@
 from functools import partial
 from typing import Optional, List
 
-from simnet.client.wish.base import BaseWishClient
+from simnet.client.components.wish.base import BaseWishClient
 from simnet.models.genshin.wish import Wish
 from simnet.utils.enum_ import Game
 from simnet.utils.paginator import WishPaginator
