+++ simnet/client/genshin.pyi
@@ -1,8 +1,8 @@
 from typing import Optional
 
-from simnet.client.account.auth import AuthClient
-from simnet.client.chronicle.genshin import GenshinChronicleClient
-from simnet.client.wish.genshin import WishClient
+from simnet.client.components.account.auth import AuthClient
+from simnet.client.components.chronicle.genshin import GenshinChronicleClient
+from simnet.client.components.wish.genshin import WishClient
 from simnet.utils.enum_ import Region
 from simnet.utils.types import CookieTypes, HeaderTypes, TimeoutTypes
